cleaning interception logic
