+++ packages/Ecotone/src/Messaging/Config/Container/MethodInterceptorsConfiguration.php
@@ -2,11 +2,8 @@
 
 namespace Ecotone\Messaging\Config\Container;
 
-use Ecotone\Messaging\Handler\InterfaceToCall;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\AroundInterceptorBuilder;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\MethodInterceptorBuilder;
-use Ecotone\Messaging\Support\Assert;
-use Ecotone\Messaging\Support\InvalidArgumentException;
 
 class MethodInterceptorsConfiguration
 {
@@ -20,69 +17,11 @@ class MethodInterceptorsConfiguration
         private array $aroundInterceptors,
         private array $afterInterceptors,
     ) {
-        self::checkInterceptorPrecedence($this->beforeInterceptors);
-        self::checkInterceptorPrecedence($this->aroundInterceptors);
-        self::checkInterceptorPrecedence($this->afterInterceptors);
-    }
-
-    private static function checkInterceptorPrecedence(array $interceptors): void
-    {
-        $lastPrecedence = null;
-        foreach ($interceptors as $interceptor) {
-            if ($lastPrecedence !== null && $interceptor->getPrecedence() < $lastPrecedence) {
-                throw InvalidArgumentException::create('Interceptors must be sorted by precedence. Found: ' . $interceptor->getPrecedence() . ' after ' . $lastPrecedence);
-            }
-            $lastPrecedence = $interceptor->getPrecedence();
-        }
     }
 
     public static function createEmpty()
     {
         return new self([], [], []);
-    }
-
-    public function getRelatedInterceptors(InterfaceToCall $interfaceToCall, array $endpointAnnotations, array $requiredInterceptorNames): self
-    {
-        return new self(
-            $this->getRelatedInterceptorsFor($this->beforeInterceptors, $interfaceToCall, $endpointAnnotations, $requiredInterceptorNames),
-            $this->getRelatedInterceptorsFor($this->aroundInterceptors, $interfaceToCall, $endpointAnnotations, $requiredInterceptorNames),
-            $this->getRelatedInterceptorsFor($this->afterInterceptors, $interfaceToCall, $endpointAnnotations, $requiredInterceptorNames),
-        );
-    }
-
-    /**
-     * @template T
-     * @param array<T> $interceptors
-     * @param InterfaceToCall $interceptedInterface
-     * @param array<AttributeDefinition> $endpointAnnotations
-     * @param array<string> $requiredInterceptorNames
-     * @return array<T>
-     * @throws \Ecotone\Messaging\MessagingException
-     */
-    private function getRelatedInterceptorsFor(array $interceptors, InterfaceToCall $interceptedInterface, array $endpointAnnotations, array $requiredInterceptorNames): iterable
-    {
-        Assert::allInstanceOfType($endpointAnnotations, AttributeDefinition::class);
-
-        $relatedInterceptors = [];
-
-        $endpointAnnotationsInstances = array_map(
-            fn (AttributeDefinition $attributeDefinition) => $attributeDefinition->instance(),
-            $endpointAnnotations
-        );
-        foreach ($interceptors as $interceptor) {
-            foreach ($requiredInterceptorNames as $requiredInterceptorName) {
-                if ($interceptor->hasName($requiredInterceptorName)) {
-                    $relatedInterceptors[] = $interceptor;
-                    break;
-                }
-            }
-
-            if ($interceptor->doesItCutWith($interceptedInterface, $endpointAnnotationsInstances)) {
-                $relatedInterceptors[] = $interceptor;
-            }
-        }
-
-        return $relatedInterceptors;
     }
 
     /**
