+++ packages/Ecotone/tests/Messaging/Unit/Handler/Processor/MethodInterceptorTest.php
@@ -2,11 +2,14 @@
 
 namespace Test\Ecotone\Messaging\Unit\Handler\Processor;
 
+use Ecotone\Messaging\Config\Container\Reference;
 use Ecotone\Messaging\Handler\InterfaceToCall;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\Converter\AttributeBuilder;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\Converter\ValueBuilder;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\MethodInterceptor;
+use Ecotone\Messaging\Handler\Processor\MethodInvoker\MethodInterceptorBuilder;
 use Ecotone\Messaging\Handler\ServiceActivator\ServiceActivatorBuilder;
+use Ecotone\Test\ComponentTestBuilder;
 use PHPUnit\Framework\TestCase;
 use Test\Ecotone\Messaging\Fixture\Annotation\Interceptor\InterceptorWithMultipleOptionalAttributes;
 use Test\Ecotone\Messaging\Fixture\Annotation\Interceptor\ResolvedPointcut\AroundInterceptorExample;
@@ -30,50 +33,6 @@ use Test\Ecotone\Messaging\Fixture\Behat\Calculating\PowerCalculation;
  */
 class MethodInterceptorTest extends TestCase
 {
-    public function test_adding_parameters_when_type_hinting_for_annotation_class()
-    {
-        $interceptorInterface = InterfaceToCall::create(CalculatorInterceptor::class, 'multiplyBefore');
-        $methodInterceptor = MethodInterceptor::create(
-            CalculatorInterceptor::class,
-            $interceptorInterface,
-            ServiceActivatorBuilder::create(CalculatorInterceptor::class, InterfaceToCall::create(CalculatorInterceptor::class, 'multiplyBefore')),
-            1,
-            ''
-        );
-
-        $this->assertEquals(
-            [
-                new AttributeBuilder('beforeMultiplyCalculation', new BeforeMultiplyCalculation(2), Calculator::class, 'calculate'),
-            ],
-            $methodInterceptor->addInterceptedInterfaceToCall(InterfaceToCall::create(Calculator::class, 'calculate'), [])
-                ->getInterceptingObject()
-                ->getParameterConverters()
-        );
-    }
-
-    public function test_resolving_nullable_attributes(): void
-    {
-        $interceptorInterface = InterfaceToCall::create(InterceptorWithMultipleOptionalAttributes::class, 'doSomething');
-        $methodInterceptor = MethodInterceptor::create(
-            InterceptorWithMultipleOptionalAttributes::class,
-            $interceptorInterface,
-            ServiceActivatorBuilder::create(InterceptorWithMultipleOptionalAttributes::class, $interceptorInterface),
-            1,
-            BeforeMultiplyCalculation::class . '||' . AfterMultiplyCalculation::class . '||' . PowerCalculation::class
-        );
-
-        $this->assertEquals(
-            [
-                new AttributeBuilder('beforeMultiplyCalculation', new BeforeMultiplyCalculation(2), Calculator::class, 'calculate'),
-                new AttributeBuilder('afterMultiplyCalculation', new AfterMultiplyCalculation(2), Calculator::class, 'calculate'),
-                new ValueBuilder('powerCalculation', null),
-            ],
-            $methodInterceptor->addInterceptedInterfaceToCall(InterfaceToCall::create(Calculator::class, 'calculate'), [])
-                ->getInterceptingObject()
-                ->getParameterConverters()
-        );
-    }
-
     public function test_resolving_pointcut_automatically()
     {
         $this->assertEquals(
